Predict Tree Surrogate

Predict the response for newdata of a TreeSurrogate object.

This function makes the TreeSurrogate object call its internal $predict() method.

# S3 method for TreeSurrogate
predict(object, newdata, type = "prob", ...)

Arguments

object
The surrogate tree. A TreeSurrogate object.

newdata
A data.frame for which to predict.

type
Either "prob" or "class". Ignored if the surrogate tree does regression.

...
Further arguments for predict_party.

Value

A data.frame with the predicted outcome. In case of regression it is the predicted \(\hat{y}\). In case of classification it is either the class probabilities (for type "prob") or the class label (type "class")

See also

Examples

library("randomForest")
# Fit a Random Forest on the Boston housing data set
data("Boston", package = "MASS")
rf <- randomForest(medv ~ ., data = Boston, ntree = 50)
# Create a model object
mod <- Predictor$new(rf, data = Boston[-which(names(Boston) == "medv")])

# Fit a decision tree as a surrogate for the whole random forest
dt <- TreeSurrogate$new(mod)

# Plot the resulting leaf nodes
predict(dt, newdata = Boston)

#> Warning: Dropping additional columns: medv

#>       .y.hat
#> 1   28.46813
#> 2   21.74514
#> 3   28.46813
#> 4   28.46813
#> 5   28.46813
#> 6   28.46813
#> 7   21.74514
#> 8   14.82292
#> 9   14.82292
#> 10  14.82292
#> 11  14.82292
#> 12  21.74514
#> 13  14.82292
#> 14  21.74514
#> 15  21.74514
#> 16  21.74514
#> 17  28.46813
#> 18  21.74514
#> 19  21.74514
#> 20  21.74514
#> 21  14.82292
#> 22  21.74514
#> 23  14.82292
#> 24  14.82292
#> 25  14.82292
#> 26  14.82292
#> 27  21.74514
#> 28  14.82292
#> 29  21.74514
#> 30  21.74514
#> 31  14.82292
#> 32  21.74514
#> 33  14.82292
#> 34  14.82292
#> 35  14.82292
#> 36  21.74514
#> 37  21.74514
#> 38  21.74514
#> 39  21.74514
#> 40  28.46813
#> 41  28.46813
#> 42  28.46813
#> 43  28.46813
#> 44  28.46813
#> 45  21.74514
#> 46  21.74514
#> 47  21.74514
#> 48  14.82292
#> 49  14.82292
#> 50  14.82292
#> 51  21.74514
#> 52  21.74514
#> 53  28.46813
#> 54  21.74514
#> 55  21.74514
#> 56  28.46813
#> 57  28.46813
#> 58  28.46813
#> 59  28.46813
#> 60  21.74514
#> 61  21.74514
#> 62  21.74514
#> 63  28.46813
#> 64  21.74514
#> 65  21.74514
#> 66  28.46813
#> 67  21.74514
#> 68  21.74514
#> 69  21.74514
#> 70  21.74514
#> 71  28.46813
#> 72  21.74514
#> 73  28.46813
#> 74  28.46813
#> 75  28.46813
#> 76  21.74514
#> 77  21.74514
#> 78  21.74514
#> 79  21.74514
#> 80  21.74514
#> 81  28.46813
#> 82  28.46813
#> 83  28.46813
#> 84  28.46813
#> 85  21.74514
#> 86  28.46813
#> 87  21.74514
#> 88  21.74514
#> 89  28.46813
#> 90  28.46813
#> 91  21.74514
#> 92  21.74514
#> 93  21.74514
#> 94  28.46813
#> 95  21.74514
#> 96  28.46813
#> 97  21.74514
#> 98  44.17342
#> 99  44.17342
#> 100 28.46813
#> 101 21.74514
#> 102 28.46813
#> 103 21.74514
#> 104 21.74514
#> 105 21.74514
#> 106 14.82292
#> 107 14.82292
#> 108 21.74514
#> 109 21.74514
#> 110 14.82292
#> 111 21.74514
#> 112 21.74514
#> 113 14.82292
#> 114 14.82292
#> 115 21.74514
#> 116 14.82292
#> 117 21.74514
#> 118 21.74514
#> 119 14.82292
#> 120 21.74514
#> 121 21.74514
#> 122 21.74514
#> 123 14.82292
#> 124 14.82292
#> 125 14.82292
#> 126 21.74514
#> 127 14.82292
#> 128 14.82292
#> 129 14.82292
#> 130 14.82292
#> 131 21.74514
#> 132 21.74514
#> 133 21.74514
#> 134 14.82292
#> 135 14.82292
#> 136 14.82292
#> 137 14.82292
#> 138 21.74514
#> 139 14.82292
#> 140 14.82292
#> 141 14.82292
#> 142 14.82292
#> 143 14.82292
#> 144 14.82292
#> 145 14.82292
#> 146 14.82292
#> 147 14.82292
#> 148 14.82292
#> 149 14.82292
#> 150 14.82292
#> 151 21.74514
#> 152 21.74514
#> 153 21.74514
#> 154 14.82292
#> 155 14.82292
#> 156 14.82292
#> 157 14.82292
#> 158 28.46813
#> 159 28.46813
#> 160 28.46813
#> 161 28.46813
#> 162 44.17342
#> 163 44.17342
#> 164 44.17342
#> 165 21.74514
#> 166 21.74514
#> 167 44.17342
#> 168 21.74514
#> 169 21.74514
#> 170 21.74514
#> 171 21.74514
#> 172 21.74514
#> 173 21.74514
#> 174 21.74514
#> 175 21.74514
#> 176 28.46813
#> 177 21.74514
#> 178 28.46813
#> 179 28.46813
#> 180 28.46813
#> 181 44.17342
#> 182 21.74514
#> 183 28.46813
#> 184 28.46813
#> 185 21.74514
#> 186 21.74514
#> 187 44.17342
#> 188 28.46813
#> 189 28.46813
#> 190 28.46813
#> 191 28.46813
#> 192 28.46813
#> 193 28.46813
#> 194 28.46813
#> 195 28.46813
#> 196 44.17342
#> 197 28.46813
#> 198 21.74514
#> 199 28.46813
#> 200 28.46813
#> 201 28.46813
#> 202 28.46813
#> 203 44.17342
#> 204 44.17342
#> 205 44.17342
#> 206 21.74514
#> 207 21.74514
#> 208 14.82292
#> 209 21.74514
#> 210 14.82292
#> 211 14.82292
#> 212 14.82292
#> 213 14.82292
#> 214 21.74514
#> 215 14.82292
#> 216 21.74514
#> 217 21.74514
#> 218 21.74514
#> 219 14.82292
#> 220 21.74514
#> 221 21.74514
#> 222 14.82292
#> 223 21.74514
#> 224 28.46813
#> 225 44.17342
#> 226 44.17342
#> 227 44.17342
#> 228 28.46813
#> 229 44.17342
#> 230 28.46813
#> 231 21.74514
#> 232 28.46813
#> 233 44.17342
#> 234 44.17342
#> 235 21.74514
#> 236 21.74514
#> 237 21.74514
#> 238 28.46813
#> 239 28.46813
#> 240 28.46813
#> 241 21.74514
#> 242 21.74514
#> 243 21.74514
#> 244 28.46813
#> 245 21.74514
#> 246 14.82292
#> 247 21.74514
#> 248 21.74514
#> 249 21.74514
#> 250 28.46813
#> 251 28.46813
#> 252 28.46813
#> 253 28.46813
#> 254 44.17342
#> 255 28.46813
#> 256 21.74514
#> 257 44.17342
#> 258 44.17342
#> 259 28.46813
#> 260 28.46813
#> 261 21.74514
#> 262 44.17342
#> 263 44.17342
#> 264 21.74514
#> 265 21.74514
#> 266 21.74514
#> 267 21.74514
#> 268 44.17342
#> 269 44.17342
#> 270 21.74514
#> 271 21.74514
#> 272 28.46813
#> 273 28.46813
#> 274 44.17342
#> 275 28.46813
#> 276 28.46813
#> 277 28.46813
#> 278 28.46813
#> 279 28.46813
#> 280 28.46813
#> 281 44.17342
#> 282 28.46813
#> 283 44.17342
#> 284 44.17342
#> 285 28.46813
#> 286 21.74514
#> 287 21.74514
#> 288 28.46813
#> 289 28.46813
#> 290 21.74514
#> 291 28.46813
#> 292 28.46813
#> 293 28.46813
#> 294 21.74514
#> 295 21.74514
#> 296 28.46813
#> 297 28.46813
#> 298 14.82292
#> 299 28.46813
#> 300 28.46813
#> 301 28.46813
#> 302 21.74514
#> 303 21.74514
#> 304 28.46813
#> 305 28.46813
#> 306 21.74514
#> 307 28.46813
#> 308 28.46813
#> 309 28.46813
#> 310 21.74514
#> 311 21.74514
#> 312 28.46813
#> 313 21.74514
#> 314 21.74514
#> 315 21.74514
#> 316 21.74514
#> 317 14.82292
#> 318 14.82292
#> 319 21.74514
#> 320 21.74514
#> 321 28.46813
#> 322 28.46813
#> 323 28.46813
#> 324 21.74514
#> 325 28.46813
#> 326 28.46813
#> 327 28.46813
#> 328 21.74514
#> 329 21.74514
#> 330 28.46813
#> 331 21.74514
#> 332 21.74514
#> 333 28.46813
#> 334 28.46813
#> 335 28.46813
#> 336 21.74514
#> 337 21.74514
#> 338 21.74514
#> 339 21.74514
#> 340 21.74514
#> 341 21.74514
#> 342 28.46813
#> 343 21.74514
#> 344 28.46813
#> 345 28.46813
#> 346 21.74514
#> 347 21.74514
#> 348 28.46813
#> 349 28.46813
#> 350 28.46813
#> 351 28.46813
#> 352 28.46813
#> 353 28.46813
#> 354 28.46813
#> 355 21.74514
#> 356 28.46813
#> 357 14.82292
#> 358 21.74514
#> 359 21.74514
#> 360 21.74514
#> 361 28.46813
#> 362 21.74514
#> 363 21.74514
#> 364 21.74514
#> 365 44.17342
#> 366 28.46813
#> 367 21.74514
#> 368 21.74514
#> 369 28.46813
#> 370 28.46813
#> 371 28.46813
#> 372 21.74514
#> 373 21.74514
#> 374 14.82292
#> 375 14.82292
#> 376 21.74514
#> 377 14.82292
#> 378 14.82292
#> 379 14.82292
#> 380 14.82292
#> 381 14.82292
#> 382 14.82292
#> 383 14.82292
#> 384 14.82292
#> 385 14.82292
#> 386 14.82292
#> 387 14.82292
#> 388 14.82292
#> 389 14.82292
#> 390 14.82292
#> 391 14.82292
#> 392 14.82292
#> 393 14.82292
#> 394 14.82292
#> 395 14.82292
#> 396 14.82292
#> 397 14.82292
#> 398 14.82292
#> 399 14.82292
#> 400 14.82292
#> 401 14.82292
#> 402 14.82292
#> 403 14.82292
#> 404 14.82292
#> 405 14.82292
#> 406 14.82292
#> 407 14.82292
#> 408 21.74514
#> 409 14.82292
#> 410 14.82292
#> 411 21.74514
#> 412 14.82292
#> 413 14.82292
#> 414 14.82292
#> 415 14.82292
#> 416 14.82292
#> 417 14.82292
#> 418 14.82292
#> 419 14.82292
#> 420 14.82292
#> 421 14.82292
#> 422 14.82292
#> 423 21.74514
#> 424 14.82292
#> 425 14.82292
#> 426 14.82292
#> 427 14.82292
#> 428 21.74514
#> 429 14.82292
#> 430 14.82292
#> 431 14.82292
#> 432 14.82292
#> 433 21.74514
#> 434 14.82292
#> 435 14.82292
#> 436 14.82292
#> 437 14.82292
#> 438 14.82292
#> 439 14.82292
#> 440 14.82292
#> 441 14.82292
#> 442 14.82292
#> 443 14.82292
#> 444 14.82292
#> 445 14.82292
#> 446 14.82292
#> 447 14.82292
#> 448 14.82292
#> 449 14.82292
#> 450 14.82292
#> 451 14.82292
#> 452 14.82292
#> 453 14.82292
#> 454 14.82292
#> 455 14.82292
#> 456 14.82292
#> 457 14.82292
#> 458 14.82292
#> 459 14.82292
#> 460 21.74514
#> 461 14.82292
#> 462 21.74514
#> 463 21.74514
#> 464 21.74514
#> 465 21.74514
#> 466 21.74514
#> 467 14.82292
#> 468 14.82292
#> 469 14.82292
#> 470 21.74514
#> 471 14.82292
#> 472 21.74514
#> 473 21.74514
#> 474 21.74514
#> 475 14.82292
#> 476 14.82292
#> 477 14.82292
#> 478 14.82292
#> 479 14.82292
#> 480 21.74514
#> 481 21.74514
#> 482 28.46813
#> 483 28.46813
#> 484 21.74514
#> 485 21.74514
#> 486 21.74514
#> 487 21.74514
#> 488 21.74514
#> 489 14.82292
#> 490 14.82292
#> 491 14.82292
#> 492 14.82292
#> 493 21.74514
#> 494 21.74514
#> 495 21.74514
#> 496 14.82292
#> 497 14.82292
#> 498 21.74514
#> 499 21.74514
#> 500 14.82292
#> 501 21.74514
#> 502 21.74514
#> 503 21.74514
#> 504 28.46813
#> 505 28.46813
#> 506 21.74514