Effect of one or two feature(s) on the model predictions (deprecated)

Effect of one or two feature(s) on the model predictions (deprecated)

Effect of one or two feature(s) on the model predictions (deprecated)

Details

Deprecated, please use FeatureEffect.

See also

Super classes

iml::InterpretationMethod -> iml::FeatureEffect -> Partial

Methods

Public methods

Partial$new()

Partial$clone()

Inherited methods

iml::InterpretationMethod$plot()

iml::InterpretationMethod$print()

iml::FeatureEffect$center()

iml::FeatureEffect$predict()

iml::FeatureEffect$set.feature()

Method new()

Effect of one or two feature(s) on the model predictions

Usage

Partial$new(
  predictor,
  feature,
  aggregation = "pdp",
  ice = TRUE,
  center.at = NULL,
  grid.size = 20
)

Arguments

predictor

Predictor
The object (created with Predictor$new()) holding the machine learning model and the data.

feature

(character(1) | character(2) | numeric(1) | numeric(2))
The feature name or index for which to compute the effects.

aggregation

(character(1))
The aggregation approach to use. Possible values are "pdp", "ale" or "none".

ice

logical
Whether to compute ice plots.

center.at

(numeric(1))
Value at which the plot should be centered. Ignored in the case of two features.

grid.size

(numeric(1) | numeric(2))
The size of the grid for evaluating the predictions.

Method clone()

The objects of this class are cloneable with this method.

Usage

Partial$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.